parsing_1/docs: replace pip's private requests with the real requests
Pip's `_vendor` module contains a copy of the `requests` library.
However, as indicated by the underscore at the start of the package
name, this is a private implementation detail that can be removed at
any time. This commit replaces Pip's vendored `requests` library with
the actual `requests` library and adds instructions on how to download
it.